requirements.txt
fastapi
uvicorn
celery
pika
kafka-python
python-dotenv
